// src/app/types/globalMetrics.ts
export interface GlobalMetrics {
  btc_dominance: number
  eth_dominance: number
  active_cryptocurrencies: number
  total_cryptocurrencies: number
  active_market_pairs: number
  active_exchanges: number
  total_exchanges: number
  last_updated: string
  total_market_cap: number
  total_volume_24h: number
  altcoin_volume_24h: number
  defi_volume_24h: number
  stablecoin_volume_24h: number
  total_market_cap_yesterday: number
  total_volume_24h_yesterday: number
  total_market_cap_yesterday_percentage_change: number
  total_volume_24h_yesterday_percentage_change: number
}